+++ src/blocks/CTA/schema.ts
@@ -1,8 +1,7 @@
-import {Theme, validators} from '@gravity-ui/page-constructor';
+import {validators} from '@gravity-ui/page-constructor';
 
 import {BlockType} from '../../models/common';
 import {BlogBlockBase} from '../../schema/common';
-import {DEFAULT_THEME} from '../../constants';
 
 const {
     subBlocks: {ContentBase},
@@ -27,30 +26,6 @@ export const CTA = {
                         ...ContentBase,
                     },
                 },
-            },
-            // fix after withTheme is added to PC exports https://github.com/gravity-ui/page-constructor/issues/1119
-            cardBackgroundColor: {
-                oneOf: [
-                    {
-                        type: 'string',
-                        optionName: 'no theme',
-                    },
-                    {
-                        type: 'object',
-                        additionalProperties: false,
-                        required: DEFAULT_THEME,
-                        properties: Object.values(Theme).reduce(
-                            (result, themeName) => ({
-                                ...result,
-                                [themeName]: {
-                                    type: 'string',
-                                },
-                            }),
-                            {},
-                        ),
-                        optionName: 'themes',
-                    },
-                ],
             },
         },
     },