+++ src/models/blocks.ts
@@ -6,7 +6,6 @@ import {
     HeaderBlockProps,
     MediaProps as PCMediaProps,
     TextTheme,
-    ThemeSupporting,
 } from '@gravity-ui/page-constructor';
 
 import {BlockType, ClassNameProps, PostData, QAProps} from './common';
@@ -37,7 +36,6 @@ export type ColoredTextProps = ContentBlockProps &
 
 export type CTAProps = QAProps & {
     items: Array<ContentBlockProps>;
-    cardBackgroundColor?: ThemeSupporting<string>;
 } & PaddingsYFMProps;
 
 export type HeaderProps = HeaderBlockProps & PaddingsYFMProps;